requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
ansicolors==1.1.8
anyio==4.10.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
contourpy @ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
controlnet_aux==0.0.10
cryptography==45.0.7
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
databricks-sdk==0.65.0
datasets==4.0.0
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1756742011/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
diffusers==0.30.3
dill @ file:///C:/b/abs_f79eg27d2q/croot/dill_1715094735295/work
docker==7.1.0
einops==0.8.1
entrypoints==0.4
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
fastapi==0.116.1
fastjsonschema==2.21.2
filelock==3.19.1
Flask==3.1.2
fonttools @ file:///C:/b/abs_4crkswws2h/croot/fonttools_1737040078745/work
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-auth==2.40.3
google-genai==1.36.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1756474536/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib @ file:///C:/b/abs_0ahqdkwgvn/croot/joblib_1754310214981/work
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
jupyterlab_widgets==3.0.15
kiwisolver @ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
lazy_loader==0.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mkl-service==2.4.0
mkl_fft @ file:///C:/b/abs_64riv3lfyn/croot/mkl_fft_1751310924052/work
mkl_random @ file:///C:/b/abs_47jvsmau4j/croot/mkl_random_1751310922820/work
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.9.0
nbclient==0.10.2
nbformat==5.10.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandarallel @ file:///home/conda/feedstock_root/build_artifacts/pandarallel_1735835625496/work
pandas @ file:///D:/bld/pandas_1755778903675/work
papermill==2.6.0
parallel-pandas==0.7
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
peft==0.17.1
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
plotly==6.3.1
ply==3.11
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
propcache==0.3.2
protobuf==6.32.1
psutil @ file:///D:/bld/psutil_1755851298735/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pyparsing @ file:///C:/b/abs_40z8gyj9wi/croot/pyparsing_1731445739241/work
PyQt5==5.15.10
PyQt5_sip @ file:///C:/b/abs_d1i1jr9fgt/croot/pyqt-split_1736540551157/work/pyqt_sip
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487145/work
PyYAML==6.0.2
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1756136275/work
referencing==0.37.0
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn @ file:///C:/b/abs_45qemhn4lg/croot/scikit-learn_1753427401078/work
scipy @ file:///C:/b/abs_a93bbipqh2/croot/scipy_1755766520463/work/dist/scipy-1.16.0-cp311-cp311-win_amd64.whl#sha256=ee46df25f27b3082dbaf7f5979c5aff19d61b830e9ab70535f26c5f9614310a4
seaborn @ file:///C:/b/abs_0bv5x0ucxb/croot/seaborn_1749110315600/work
sip @ file:///C:/b/abs_b6twfwtueg/croot/sip_1736544693649/work
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl @ file:///C:/b/abs_def0dwqlft/croot/threadpoolctl_1719407816649/work
tifffile==2025.8.28
timm==1.0.19
tokenizers==0.19.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado @ file:///D:/bld/tornado_1756854822157/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.44.2
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
unicodedata2 @ file:///C:/b/abs_dfnftvxi4k/croot/unicodedata2_1736543771112/work
urllib3==2.5.0
uvicorn==0.35.0
waitress==3.0.2
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work

